refactor(ui): split toolkit detail into a tabbed safety console
The 792-line ToolkitDetailBody god component becomes a thin tabbed shell
(Overview / Access / Keys / Settings, deep-linkable via ?tab= like
Monitor) with one component per tab, each owning its queries/mutations.
The kill switch, id chip, and suspended banner stay pinned above the
tabs so suspension is reachable from every section; Edit moves from a
dialog to an inline Settings form and Delete into a danger zone there.
Drops the dead layout='sheet' branch left over from the pre-#797 sheet
host.
